server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
pandas==2.1.4
python-dotenv==1.0.0
alembic==1.13.1
# Performance optimizations
redis==5.0.1
aioredis==2.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
# Monitoring and profiling
prometheus-client==0.19.0
psutil==5.9.6
# Caching
cachetools==5.3.2
# Compression
brotli==1.1.0
